Do not install licenses
Firstly, we should certainly not be installing something as generic as
standard license texts to a location as generic as share/LICENSES.

Secondly, it is not our job to install license text at all.  We provide
licenses with all parts of our source distributions, and it is up to the
producers of binary distributions to decide on the best way to make
license text available for those.

REVIEW: 114979