The outputFile parameter is not used by the stylesheets, don't pass it
This solve an old issue: if a directory starts with ".", it was passed
unquoted to libxslt which would complain with an error like:
-------------------
XPath error : Invalid expression
/home/kde-devel/.cache5/khelpcenter/help/__home__kde-
devel__kde__share__doc__HTML__en__kioslave__file__index.docbook
                     ^
runtime error
Evaluating user parameter outputFile failed
-------------------
Excluding the parameter on windows-only was just a workaround, but most
probably the issue has been on UNIX systems too for a long time.

Another way to solve the bug is quoting the _value_ of the parameter using
single quotes '...', but anyway in this case the name of output file is
not used, or I can't find any occurrence in the stylesheets.
The stylesheet is applied and the name of the file is used only after
to write the generated XML (see tranform() function).

CCBUG: 246755
REVIEW: 116604