Update install-compile.docbook translations to not reference KDE4
These are links to techbase. Link text that was already English was just
updated. Spanish was supplied by apol. Korean was grabbed from the
Korean translation of techbase. The remaining ones I just replaced
"KDE4" with "KDE Software" (capitalized like that so it is like a
coherent term). Translators can touch up the text later if they wish.

REVIEW: 117758