Always set src_dir to a not-empty value ('.')
src_dir is empty when the current directory contains the docbook the
function is working on. But the subsequent globs use its value, so
the search pattern is incorrect ('/*.docbook' instead of, after the
patch, './*.docbook'), so that docbook and png files are not installed,
breaking the help:/ kioslave.

REVIEW: 118158