No symlink to common/ (broken, now unused, UNIX only)
The common/ directory (which contains per-language common files
for documentation) has been renamed (kdoctools5-common);
even if it was not, the symlink is not needed anymore in order
to retrieve the documentation (the references to the content in
that directory are now accessed through the help:/ kioslave).

The change described above removes/solves the long-standing TODO
for non-UNIX platform(s).

REVIEW: 118546