Also find catalog files that are symlinks
When building KF5 on Mac OS X with homebrew, the catalog and
other files are symlinked from the 'Cellar' (install dir) to the
homebrew prefix. After that, meinproc5 wouldn't find its catalog
because it decides to ignore symbolic links. Removing that option
fixes the build on Mac OS X.

REVIEW: 119567
CHANGELOG: Fix the build on MacOSX when using Homebrew