Allow to disable KArchive for internal usage
Most of the code was already prepared to compile without the KArchive
dependency (including the MEINPROC_NO_KARCHIVE flag).
This mode should be only enabled on KDE infrastructure, as the compressed
cache (the only added feature coming from KArchive) is not used by the
generator of the documentation website.
A warning is printed when the mode is enabled.
The saveToCache function is preserved, even if without code, to minimize
the changes required.

CHANGELOG: Allow minimal mode (no KArchive), only for KDE infrastructure
REVIEW: 126648