Use unicode characters for callouts
Callouts use static image by default. Their path used to point
(until kdelibs4) to a local copy of images, but they have been
removed since then. docbook-xsl ships with images, but the path
should be adjusted anyway to find them.
Moreover, the definition of the parameter in kde-style.xsl shadowed
the overridden parameter in the other file which included kde-style,
so the definition fom that file should have been fixed as well.
On the other side, it is possible to use unicode characters instead,
and it is reasonable to assume that modern systems can provide
those codepoints.

The only documentations affected so far are the ones of Konversation
and Okteta.