Call kdoctools_create_handbook only for index.docbook
When more documentation files are available in a certain directory,
the assumption is that there should always be an index.docbook which
imports all the other docbook files in the same directory. The call
to kdoctools_create_handbook must be created only for index.docbook;
the macro takes care of the installation of the others docbooks.
This should make kdoctools_install work when there are multiple docbooks
in the same directory.

BUG: 357428
REVIEW: 127935