Fix test: use the documented structure for docbooks
The recognized and tested structure is:
<translation_dir>/<lang>/docs/<program_name>/*.docbook

Being inside <program_name> is not important for the code which
handles manpages.

It could be possible to handle <translation_dir>/<lang>/docs/*.docbook
as (apparently) originally planned, but I think that it would introduce
a bit more of code without too much gain (the name of the program/module
needs to be extracted somehow anyway, SUBDIR is a mandatory parameter).