kdoctools_install: match the full path for the program
Summary:
In order to get the name of the program, match against the full path
which includes the path of the directory currently scanned for a
certain language.
So if docs/ appears somewhere in the path, it does not interfere
with the matching regular expression and the name of the program
is correctly extracted.

BUG: 374435

Test Plan: kdoctools_install(l10n-docs) now works

Reviewers: #frameworks, #documentation, aacid

Reviewed By: #documentation, aacid

Tags: #frameworks, #documentation

Differential Revision: https://phabricator.kde.org/D3912