Export a proper shared library for KDocTools
Summary:
This is a unique situation, as the static library was supposed to be
temporary, and the only users are kio_help/kio_ghelp inside KIO.
The shared library is going to be the "real" supported library of
KDocTools, with more content in the future.
- prepare a new KF5::DocTools shared library with the same content,
  but export only the functions which are really used by KIO.
- still install the old library (minus an unused header) to allow
  KIO to still compile until the patches which makes it use the
  shared library is ported.
No functionality should be affected.

Test Plan: The code compiles, and KIO still compiles (kio_help apparently works).

Reviewers: dfaure

Reviewed By: dfaure

Subscribers: #frameworks, #documentation

Tags: #frameworks, #documentation

Differential Revision: https://phabricator.kde.org/D4849