Solve a visibility conflict for meinproc5
Summary:
This seems to affect only mingw, which can't link meinproc5 as
the symbol exported affects in the sources shared with the
library affects the build (but not on gcc on *nix and with MSVC).

The other (probably better) solution would have been to split the
source code so that the sources of the libraries are separate, but
that would require more changes. As this seems to work, I'd prefer
to move the more complicated changes to "later".

Thanks to Hannah Von Reth for the hint.

BUG: 379142

Test Plan: Compiles with gcc 6.3 (Debian Testing), as before), with mingw (as deployed on current Craft) and with MSVC (2017) (the latter with less linking warnings so it was an issue there). All on x86_64.

Reviewers: vonreth

Reviewed By: vonreth

Subscribers: #frameworks, #documentation

Tags: #frameworks, #documentation, #windows

Differential Revision: https://phabricator.kde.org/D5674