Workaround for MSVC and catalog loading
Summary:
Apparently setting XML_CATALOG_FILES has no effects on Windows
at least with MSVC (2017). The variable seems to be set, so
there may be an issue with the linking, or in libxml2, or
in some other unidentified place. Forcibly setting the
catalog seems to work, but let's leave it as a special case
on MSVC only for now.

Test Plan:
Tested on the new CI, tested a package with documentation
(kconfigwidgets)

Reviewers: bcooksley

Reviewed By: bcooksley

Subscribers: bcooksley, #frameworks, #documentation

Tags: #frameworks, #documentation

Differential Revision: https://phabricator.kde.org/D5725