CMake: Fix target name shortening when build dir has special characters
The DocTools CMake macros generate a target name based on the path of the
file to generate. Since the target name was too long sometimes, especially
on Windows (which has path length limits), in a624463c62 I added a regex
replacement to strip the build directory from the path before transforming
it to a target name.

However, this was causing an error if the build directory had any regex
special characters, such as '+'.

I'm now using file(RELATIVE_PATH) instead of a regular expression to strip
off the build dir root, which should be more reliable and work for any
characters.

BUG:377573