Use specified Perl interpreter instead of relying on PATH
Summary: If a PERL_EXECUTABLE is passed on the CMake command line, e.g. because the default one in PATH is insufficient, kdoctools would still use the one in PATH incorrectly.  This fixes it to use the specified perl both at configure time and at build / install time.

Reviewers: ltoscano

Reviewed By: ltoscano

Subscribers: ltoscano, vonreth, kde-doc-english, kde-frameworks-devel

Tags: #frameworks, #documentation

Differential Revision: https://phabricator.kde.org/D15927