Hook up customizations for sq (Albanian) and de-duplicate
They were added in 2011 [1] but apparently never hooked up with the
build system.
Also remove the weird duplication of KDOCTOOLS_LANGUAGES_LIST. It
isn't needed and because the "set" invocation uses a variable
reference instead of the variable name it probably creates a variable
with the name "af" containing "bg ca cs da ...".

This commit allows building the Albanian documentation of kturtle,
which currently fails due some missing entities. As such it's a
stopgap measure, likely to be replaced with a dynamic language list.

[1] 059121245d6d2b74287616d79416412680348ac0 in kdelibs.git