Remove the usage of QTextCodec (and thus Qt6Core5Compat) for Qt6
This is done by taking the current Windows code path which assumes we
always want to produce UTF-8 output. That's fine in general, rendering
UTF-8 HTML works everywhere. It's also what the Unix code path would
produce with Qt6 anyway:
- The equivalent to QTextCodec::codecForLocale() would QStringConverter's
System encoding, which is UTF-8 on Unix in Qt6.
- There is no external setCodecForLocale() call which could change this,
the only external user of this is kio_help.
- With UTF-8 the Unix code path always takes the early exit in the while
loop, making it produce the same result as the Windows code path.