meinproc: Remove dead entry parsing code.
Commit e44d403cd9f2e37816493cc8c2a354cfb6f071ef removed the only usage
of this file-local function, which I confirmed with a ripgrep search
across the repository.  Removing this made MyPair and PairList dead code
as well, and these were also removed.

Tested by building, installing, and using the resulting executable to
build kdesrc-build docs, and by ensuring the kdoctools autotests still
pass.