meinproc: Fix usage of QString reference to temporary.
Caught by clazy:
    /kf5/frameworks/kdoctools/src/meinproc.cpp:94:35:
    warning: Don't call QList::first() on temporary
    [-Wclazy-detaching-temporary]
        const QString checkFilename = parser.positionalArguments().first();

Although we know that .first() must be a valid reference by this point
in the code, it's a reference to a temporary and using this reference is
undefined behavior even though QStringList dtor is clearly not going to
do anything with the memory as it is still referenced by `parser`.

But it's undefined behavior all the same and who knows what interesting
optimizations the compiler will cook up for this someday if left
unfixed. Fixed by making the temporary a named variable.