kdoctools_install: fix doc detection in path with special chars
${lang_dir} represents the absolute path of a directory within the
sources that contains the files for a language; if this path contains
special characters (e.g. "+"), then using it in a regex such as
  "${lang_dir}/docs/(.*)/index.docbook"
will not produce matches.

To avoid special characters getting in the way of the lookup for
manpages or index.docbook files, restrict the results of the glob search
relative to ${lang_dir}, making sure to compose full paths back when
invoking kdoctools_create_manpage() and kdoctools_create_handbook().

To make sure it can be tested and it does not regress, rename the
existing "kdoctools-install" autotest to "kdoctools+install".

(cherry picked from commit b3cc46ca3d643ce88d984daf0148ecc70ae050eb)