Provide option to enable relocatable docbook files on non WIN32 platforms
Using relocatable docbook files should only be used when specifically
needed by the build system or by the platform. The known use cases are
- building for WIN32
- building in a cross-building setup where the native system's build
  directy might not exist anymore during target package building (target
  package then has a sysroot of its own but with files being located at
  different locations)
This change provides a RELOCATABLE_DOCBOOK_FILES option that can be
enabled when needed in a certain setup and prepends the CMAKE_SYSROOT
variable (only to be set for cross-compilation).